feat: move work experiences validation schemas to a new file for better organization
