+++ backend/src/db/schema/workExperiences.ts
@@ -49,20 +49,3 @@ export const workExperiencesRelations = relations(
     }),
   }),
 );
-
-// Zod schemas for validation
-export const selectWorkExperiencesSchema = createSelectSchema(workExperiences);
-export const insertWorkExperiencesSchema = createInsertSchema(workExperiences, {
-  companyName: z.string().min(1, "Company name is required").max(100),
-  current: z.boolean().default(false),
-  startDate: z.iso.datetime(),
-  endDate: z.iso.datetime().optional(),
-});
-export const updateWorkExperiencesSchema = createUpdateSchema(
-  workExperiences,
-).omit({ userProfileId: true });
-
-// Type exports
-export type WorkExperience = z.infer<typeof selectWorkExperiencesSchema>;
-export type UpdateWorkExperience = z.infer<typeof updateWorkExperiencesSchema>;
-export type InsertWorkExperience = z.infer<typeof insertWorkExperiencesSchema>;
